Implementing Lab2
Lab2 does require the base logic from Lab1, so did a pretty big refactor to how the stream processing happens so that it's now easier to work with.

We always add the FetchSalesData tool, as each lab needs it, there's now another method that can add additional tools that the lab will require. This also saw the need to bring the handler up for FetchSalesData, so instead it's now providing another method that can be called if you have another tool to call